# server/swagger.yml
openapi: "3.0.3"
info:
  title: SugarBox API by arairon
  version: 0.1.0
  description: A simple storage for SugarCube saves.
  license:
    name: GPL-3.0
    url: https://opensource.org/license/gpl-3-0
  contact:
    name: arai.icu
    url: https://arai.icu
    email: arai@arai.icu
servers:
  - url: https://sugarbox.arai.icu/api
  #- url: http://127.0.0.1:3000/api
paths:
...
  /saves/new:
    post:
      summary: Create a new save
      description: Creates a new save for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCreateSchema"
      responses:
        "200":
          description: Successfully created a new save
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveSchema"
        "400":
          description: Validation error
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /saves/uuid/{saveId}:
    get:
      summary: Get a save by UUID
      description: Retrieves a specific save by its UUID for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: charId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the save to retrieve.
      responses:
        "200":
          description: Successfully retrieved the save
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveSchema"
        "400":
          description: Invalid UUID
        "404":
          description: Character not found
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
    patch:
      summary: Update a save by UUID
      description: Updates a specific save by its UUID for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: saveId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the save to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveUpdateSchema"
      responses:
        "200":
          description: Successfully updated the save
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveSchema"
        "400":
          description: Invalid UUID or validation error
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error

    delete:
      summary: Delete a save by UUID
      description: Deletes a specific save by its UUID for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: saveId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the save to delete.
      responses:
        "200":
          description: Successfully deleted the save
        "400":
          description: Invalid UUID
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
...
components:
  schemas:
...
    SaveSchema:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the save.
        uuid:
          type: string
          format: uuid
          description: UUID of the save.
        name:
          type: string
          maxLength: 100
          description: Name of the save.
        description:
          type: string
          maxLength: 256
          description: Description of the save.
        gameVersion:
          type: string
          description: Version of the game associated with the save.
        gameId:
          type: string
          format: uuid
          description: UUID of the game associated with the save.
        charId:
          type: string
          format: uuid
          description: UUID of the character associated with the save.
        ownerId:
          type: integer
          description: ID of the owner of the save.
        data:
          type: string
          maxLength: 1000000
          description: Save data (up to 1MB).
        size:
          type: integer
          description: Size of the save data in bytes.
        hash:
          type: string
          maxLength: 256
          description: Hash of the save data.
        archived:
          type: boolean
          description: Whether the save is archived.
        archivedAt:
          type: string
          format: date-time
          description: Timestamp when the save was archived.
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the save was last updated.
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the save was created.
    SaveCreateSchema:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          description: UUID of the save.
        name:
          type: string
          maxLength: 100
          description: Name of the save.
        description:
          type: string
          maxLength: 256
          description: Description of the save.
        gameVersion:
          type: string
          description: Version of the game associated with the save.
        gameId:
          type: string
          format: uuid
          description: UUID of the game associated with the save.
        charId:
          type: string
          format: uuid
          description: UUID of the character associated with the save.
        data:
          type: string
          maxLength: 1000000
          description: Save data (up to 1MB).
        size:
          type: integer
          description: Size of the save data in bytes.
        hash:
          type: string
          maxLength: 256
          description: Hash of the save data.
    SaveUpdateSchema:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: Name of the save.
        description:
          type: string
          maxLength: 256
          description: Description of the save.
        gameVersion:
          type: string
          description: Version of the game associated with the save.
        gameId:
          type: string
          format: uuid
          description: UUID of the game associated with the save.
        charId:
          type: string
          format: uuid
          description: UUID of the character associated with the save.
        data:
          type: string
          maxLength: 1000000
          description: Save data (optional, up to 1MB).
        archived:
          type: boolean
          description: Whether the save is archived.
...